// AMM/Izumi/types.ts
export type Token = {
    chainId: number
    address: string
    name: string
    symbol: string
    decimals: number
    logoURI: string
}

export type Pool = {
    tokenA: Token
    tokenB: Token
    pair: string
    fees: number
    tickSpacing: number,
    liquidity: bigint,
    sqrtPriceX96: bigint,
    tick: number,
    Quoter: Contract
    Pool: Contract
}
// ...
export enum Fees {
    VERY_LOW = 100, // 0.01%
    LOW = 500,      // 0.05%
    MEDIUM = 3000,  // 0.3%
    BIG = 10000     // 1%
}
// ...
export type Chains = 'arbitrum' | 'polygon' | 'optimism' | 'ethereum' | 'bsc' | 'polygonTestnet' | 'arbitrumTestnet' |
                    'ethereumTestnet' | 'optimismTestnet'